Refactor wagmi utils
